% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/chartJSRadar.R
\name{chartJSRadar}
\alias{chartJSRadar}
\title{Make a ChartJS Radar Plot}
\usage{
chartJSRadar(scores, labs, width = NULL, height = NULL, maxScale = NULL,
  scaleStepWidth = NULL, scaleStartValue = 0, responsive = TRUE,
  labelSize = 18, addDots = TRUE, colMatrix = diag(3) * 255,
  polyAlpha = 0.2, lineAlpha = 0.8, ...)
}
\arguments{
\item{scores}{Named list of scores for each axis}

\item{labs}{Labels for each axis}

\item{width}{Width of output plot}

\item{height}{Height of output plot}

\item{maxScale}{Max value on each axis}

\item{scaleStepWidth}{Spacing between rings on radar}

\item{scaleStartValue}{Value at the centre of the radar}

\item{responsive}{Logical. whether or not the chart should be responsive and resize when the browser does}

\item{labelSize}{Numeric. Point label font size in pixels}

\item{addDots}{Logical. Whether to show a dot for each point}

\item{colMatrix}{Numeric matrix of rgb colour values}

\item{polyAlpha}{Alpha value for the fill of polygons}

\item{lineAlpha}{Alpha value for the outlines}

\item{...}{Extra options passed straight to chart.js}
}
\description{
R bindings to the radar plot in the chartJS library
}
\examples{
labs <- c("Communicator", "Data Wangler", "Programmer", "Technologist",  "Modeller", "Visualizer")

scores <- list("Rich" = c(9, 7, 4, 5, 3, 7),
 "Andy" = c(7, 6, 6, 2, 6, 9),
 "Aimee" = c(6, 5, 8, 4, 7, 6))

# Default settings
chartJSRadar(scores=scores, labs=labs)

# Fix the max score
chartJSRadar(scores=scores, labs=labs, maxScale=10)

# Fix max and spacing
chartJSRadar(scores=scores, labs=labs, maxScale=12, scaleStepWidth = 2)

# Add pass through settings for extra options
chartJSRadar(scores=scores, labs=labs, maxScale =10, scaleLineWidth=5)
}

