% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chartJSRadar.R
\name{chartJSRadar_html}
\alias{chartJSRadar_html}
\title{Tell htmltools where to output the chart}
\usage{
chartJSRadar_html(id, style, class, width, height, ...)
}
\arguments{
\item{id}{The id of the target object}

\item{style}{css stylings}

\item{class}{class of the target}

\item{width}{width of target}

\item{height}{height of target}

\item{...}{extra arguments currently unused}
}
\description{
Tell htmltools where to output the chart
}

