% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colourMatrix.R
\name{colourMatrix}
\alias{colourMatrix}
\title{Check and prep the colour matrix}
\usage{
colourMatrix(colMatrix)
}
\arguments{
\item{colMatrix}{A 3 x n matrix of integers between 0-255}
}
\value{
The checked and prepped matrix of the same size
}
\description{
Check and prep the colour matrix
}
\examples{
radarchart:::colourMatrix(diag(255, nrow=3))

}

