% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hello2.R
\name{hello2Output}
\alias{hello2Output}
\title{Widget output function for use in Shiny}
\usage{
hello2Output(outputId, width = "100\%", height = "400px")
}
\description{
Widget output function for use in Shiny
}

