% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hello2.R
\name{renderHello2}
\alias{renderHello2}
\title{Widget render function for use in Shiny}
\usage{
renderHello2(expr, env = parent.frame(), quoted = FALSE)
}
\description{
Widget render function for use in Shiny
}

