% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runExample.R
\name{runExampleApp}
\alias{runExampleApp}
\title{Run an example Shiny app}
\usage{
runExampleApp(example)
}
\arguments{
\item{example}{the name of the example. Choose from "basic" or "options".}
}
\description{
The radarchart package contains a number of demo Shiny apps to illustrate how
to use the plots. The code is in \code{inst/shiny-examples/} and running this
function will allow quick access to the apps.
}
\examples{
\dontrun{
runExample("basic")
}
}

