% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chartJSRadar.R
\name{setRadarScale}
\alias{setRadarScale}
\title{Autoscale the radar plot}
\usage{
setRadarScale(maxScale = NULL, scaleStepWidth = NULL, scaleStartValue = 0)
}
\arguments{
\item{maxScale}{Numeric length 1. Desired max limit}

\item{scaleStepWidth}{Numeric length 1. Spacing between rings}

\item{scaleStartValue}{Numeric length 1. Value of the centre}
}
\value{
A list containing the scale options for chartjs
}
\description{
Autoscale the radar plot
}
\examples{
\dontrun{
setRadarScale(15, 3)
setRadarScale(15, 5, 2)
}
}

