% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{skills}
\alias{skills}
\title{Skills in a team}
\format{A data frame with 6 rows and 4 columns
\describe{
  \item{Label}{The axis label for chartJSRadar}
  \item{Aimee}{Vector of skills for Aimee}
  \item{Andy}{Vector of skills for Andy}
  \item{Rich}{Vector of skills for Rich}
}}
\source{
Simulated
}
\usage{
skills
}
\description{
A dataset containing the skills vectors for three people
}
\keyword{datasets}

