% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{skillsByName}
\alias{skillsByName}
\title{Rotated version of skills data}
\format{A data frame with 6 rows and 4 columns
\describe{
  \item{Name}{Name of the team member}
  \item{Communicator}{Their Communicator score: 0-10}
  \item{Data Wangler}{Their Data Wangler score: 0-10}
  \item{Modeller}{Their Modeller score: 0-10}
  \item{Programmer}{Their Programmer score: 0-10}
  \item{Technologist}{Their Technologist score: 0-10}
  \item{Visualizer}{Their Visualizer score: 0-10}
}}
\source{
Simulated
}
\usage{
skillsByName
}
\description{
A dataset containing the skills vectors for three people but by row rather
than column. This data set is used to show how to rotate the data into a 
format accepted by \code{\link{chartJSRadar}}.
}
\keyword{datasets}

