% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chartJSRadar.R
\name{chartJSRadarOutput}
\alias{chartJSRadarOutput}
\title{Widget output function for use in Shiny}
\usage{
chartJSRadarOutput(outputId, width = "450", height = "300")
}
\arguments{
\item{outputId}{output variable to read from}

\item{width}{Must be valid CSS unit}

\item{height}{Must be valid CSS unit}
}
\description{
Widget output function for use in Shiny
}

