% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runExample.R
\name{runExampleApp}
\alias{runExampleApp}
\title{Title}
\usage{
runExampleApp(example)
}
\arguments{
\item{example}{the name of the example. Choose from "basic" or "options"}
}
\description{
Title
}
\examples{
\dontrun{
runExample("basic")
}
}

